do not create tf-idf files if they already exist

+++ python/tfidf.py
@@ -15,13 +15,14 @@ def create_tfidf_files(business_terms_dict: dict, out_path: str) -> None:
         file_name = f"{list(business_terms_dict.keys())[doc_id]}.tfidf.csv"
         print(f"create: {file_name}")
         out_file_path = os.path.join(out_path, "tfidf", file_name)
-        if not os.path.exists(os.path.dirname(out_file_path)):
-            os.makedirs(os.path.dirname(out_file_path))
-        out_file = open(out_file_path, mode="w")
-        for term_id in range(doc_term_matrix.shape[1]):
-            tfidf_value = doc_term_matrix[doc_id, term_id]
-            if tfidf_value > 0:
-                print(f"{feature_names_out[term_id]}\t{str(round(tfidf_value, 2))}", file=out_file)
+        if not os.path.exists(out_file_path):
+            if not os.path.exists(os.path.dirname(out_file_path)):
+                os.makedirs(os.path.dirname(out_file_path))
+            out_file = open(out_file_path, mode="w")
+            for term_id in range(doc_term_matrix.shape[1]):
+                tfidf_value = doc_term_matrix[doc_id, term_id]
+                if tfidf_value > 0:
+                    print(f"{feature_names_out[term_id]}\t{str(round(tfidf_value, 2))}", file=out_file)
 
     corpus = [[(v, doc_term_matrix[d, v]) for v in range(doc_term_matrix.shape[1]) if doc_term_matrix[d, v] > 0] for d in range(doc_term_matrix.shape[0])]
     dictionary = {t: feature_names_out[t] for t in range(len(feature_names_out))}
